pokemon example fix
