+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # CHANGELOG
 
+## 7.0.0-beta.7
+
+- pokemon example fix
+
 ## 7.0.0-beta.3
+
 - document generated in separate variable for easier usage
 
 ## 7.0.0-beta.2
+
 - package nnbd migraion
 
 ## 7.0.0-beta.1
+
 **MAJOR BREAKING CHANGE**
 
 - Code generated by Artemis is now nnbd-compliant
