+++ example/pokemon/pubspec.yaml
@@ -5,14 +5,14 @@ authors:
   - Igor Borges <igor@borges.me>
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  http: ^0.12.0+2
+  http: ^0.13.1
 
 dev_dependencies:
-  test: ^1.6.3
-  build_runner: ^1.5.0
-  json_serializable: ^3.0.0
+  test: ^1.16.8
+  build_runner: ^1.12.2
+  json_serializable: ^4.1.0
   artemis:
     path: ../../.
